img uploads cloudinary

+++ pages/api/products/[id].ts
@@ -1,6 +1,7 @@
 import dbConnect from '@utils/mongo';
 import type { NextApiRequest, NextApiResponse } from 'next';
 import Product from '@models/Product';
+import axios from 'axios';
 
 export default async function handler(
   req: NextApiRequest,
@@ -9,11 +10,16 @@ export default async function handler(
   const {
     method,
     query: { id },
+    cookies,
   } = req;
+  const token = cookies.token;
 
   await dbConnect();
 
   if (method === 'PUT') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json('admin이 아닙니다');
+    }
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
         new: true,
@@ -25,6 +31,9 @@ export default async function handler(
   }
 
   if (method === 'DELETE') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json('admin이 아닙니다');
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json('The product has been deleted!');
